+++ finagle-core/src/test/scala/com/twitter/finagle/loadbalancer/aperture/ApertureSuite.scala
@@ -44,7 +44,7 @@ private[loadbalancer] trait ApertureSuite {
 
   case class Factory(i: Int) extends EndpointFactory[Unit, Unit] {
     def remake() = {}
-    def address: Address = Address.Failed(new Exception)
+    val address: Address = Address.Failed(new Exception)
 
     var _total = 0
     var _outstanding = 0