+++ finagle-core/src/test/scala/com/twitter/finagle/loadbalancer/roundrobin/RoundRobinSuite.scala
@@ -1,9 +1,12 @@
 package com.twitter.finagle.loadbalancer.roundrobin
 
-import com.twitter.finagle.stats.{NullStatsReceiver, StatsReceiver}
-import com.twitter.finagle.{Address, NoBrokersAvailableException}
+import com.twitter.finagle.stats.NullStatsReceiver
+import com.twitter.finagle.stats.StatsReceiver
+import com.twitter.finagle.Address
+import com.twitter.finagle.NoBrokersAvailableException
 import com.twitter.finagle.loadbalancer.EndpointFactory
-import com.twitter.util.{Activity, Var}
+import com.twitter.util.Activity
+import com.twitter.util.Var
 
 trait RoundRobinSuite {
   // number of servers
@@ -15,7 +18,7 @@ trait RoundRobinSuite {
 
   trait RRServiceFactory extends EndpointFactory[Unit, Int] {
     def remake() = {}
-    def address = Address.Failed(new Exception)
+    val address = Address.Failed(new Exception)
     def meanLoad: Double
   }